// 2T/REACT/mi_porfolio - copia/src/model/interfaces/formacion/IFormacion.ts
export interface IFormacion {
    id:                number;
    tipo:              string;
    fecha_inicio:      string;
    fecha_fin:         string;
    nombre:            string;
    descripcion:       string;
    imagen:            string;
    institucion:       string;
    lugar_institucion: string;
}